+++ geo/CHANGES.md
@@ -28,6 +28,8 @@
   * <https://github.com/georust/geo/pull/1091>
 * Add `wkt!` macro to define geometries at compile time.
   <https://github.com/georust/geo/pull/1063>
+* Add `TriangulateDelaunay` and `VoronoiDiagram` traits.
+  * <https://github.com/franklin-ai/geo/pull/2>
 
 ## 0.26.0